Less flaky data access limits
Genoverse habitually requests more data than actually shown, which meant
that towards the 100k limit it would request a lot more than that, and
then get only the genes (but still be expecting transcript data).

Also it doesn't request data again at higher resolution when zooming in
(even with additional, different parameters attached), so if you started
at a wide zoom and then zoomed in you never gained the missing
transcript-level data.

So: implement a dataRequestLimit, which limits the amount of data
requested in a single request (by sending multiple requests as
necessary) on the client, and then implement the server limit on
requests above that amount. Which keeps the requests reasonably-sized.